requirements.txt
django
djangorestframework
psycopg
django-stubs
Pillow